Change the type data for type data

+++ src/main/java/com/nameless/lab4/dto/PointResponseDTO.java
@@ -30,5 +30,5 @@ public class PointResponseDTO {
     private String currentTime;
 
     @NotNull
-    private Integer executionTime;
+    private Long executionTime;
 }
